# manage/reference/swagger-vmrest.json
{
   "swagger": "2.0",
   "info": {
      "title": "VMware Workstation REST API",
      "version": "1.3.1",
      "description": "vmrest 1.3.1 25388281",
      "termsOfService": "http://swagger.io/terms/"
   },
   "securityDefinitions": {
      "BasicAuth": {
         "type": "basic"
      }
   },
   "basePath": "/api",
   "schemes": [
      "http",
      "https"
   ],
...
   "consumes": [
      "application/vnd.vmware.vmw.rest-v1+json"
   ],
   "produces": [
      "application/vnd.vmware.vmw.rest-v1+json"
   ],
   "paths": {
      "/vms": {
         "get": {
           "tags": [
               "VM Management"
            ],
            "summary": "Returns a list of VM IDs and paths for all VMs",
            "operationId": "getAllVMs",
            "responses": {
               "200": {
                  "description": "All VMs' ID and path",
                  "schema": {
                     "type": "array",
                     "items": {
                        "$ref": "#/definitions/VMID"
                     }
                  }
               },
               "401": {
                  "description": "Authentication failed",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "403": {
                  "description": "Permission denied",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "500": {
                  "description": "Server error",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "default": {
                  "description": "Unexpected error",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               }
            }
         },
         "post": {
           "tags": [
               "VM Management"
            ],
            "summary": "Creates a copy of the VM",
            "operationId": "createVM",
            "parameters": [
               {
                  "name": "params",
                  "in": "body",
                  "description": "Parameters of VM to create",
                  "required": true,
                  "schema": {
                     "$ref": "#/definitions/VMCloneParameter"
                  }
               },
               {
                 "name": "vmPassword",
                  "in": "query",
                  "description": "VM password",
                  "required": false,
                  "type": "string",
                  "format": "password"
               }
            ],
            "responses": {
               "201": {
                  "description": "VM information",
                  "schema": {
                     "$ref": "#/definitions/VMInformation"
                  }
               },
               "400": {
                  "description": "Invalid parameters",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "401": {
                  "description": "Authentication failed",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "403": {
                  "description": "Permission denied",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "404": {
                  "description": "No such resource",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "406": {
                  "description": "Content type was not supported",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "409": {
                  "description": "Resource state conflicts",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "500": {
                  "description": "Server error",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "default": {
                  "description": "Unexpected error",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               }
            }
         }
      },
...
      "/vms/{id}/power": {
         "get": {
            "tags": [
               "VM Power Management"
            ],
            "summary": "Returns the power state of the VM",
            "operationId": "getPowerState",
            "parameters": [
               {
                  "name": "id",
                  "in": "path",
                  "description": "ID of VM",
                  "required": true,
                  "type": "string"
               },
               {
                 "name": "vmPassword",
                  "in": "query",
                  "description": "VM password for encrypted VM. If VM is already powered on password is not needed.",
                  "required": false,
                  "type": "string",
                  "format": "password"
               }
            ],
            "responses": {
               "200": {
                  "description": "VM power state",
                  "schema": {
                     "$ref": "#/definitions/VMPowerState"
                  }
               },
               "400": {
                  "description": "Invalid parameters",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "401": {
                  "description": "Authentication failed",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "403": {
                  "description": "Permission denied",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "404": {
                  "description": "No such resource",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "500": {
                  "description": "Server error",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "default": {
                  "description": "Unexpected error",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               }
            }
         },
         "put": {
            "tags": [
               "VM Power Management"
            ],
            "summary": "Changes the VM power state",
            "operationId": "changePowerState",
            "parameters": [
               {
                  "name": "id",
                  "in": "path",
                  "description": "ID of VM",
                  "required": true,
                  "type": "string"
               },
               {
                  "name": "operation",
                  "in": "body",
                  "description": "VM power operation: on, off, shutdown, suspend, pause, unpause",
                  "required": true,
                  "schema": {
                     "$ref": "#/definitions/VMPowerOperation"
                  }
               },
               {
                 "name": "vmPassword",
                  "in": "query",
                  "description": "VM password for encrypted VM. If VM is already powered on password is not needed.",
                  "required": false,
                  "type": "string",
                  "format": "password"
               }
            ],
            "responses": {
               "200": {
                  "description": "VM power state",
                  "schema": {
                     "$ref": "#/definitions/VMPowerState"
                  }
               },
               "400": {
                  "description": "Invalid parameters",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "401": {
                  "description": "Authentication failed",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "403": {
                  "description": "Permission denied",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "404": {
                  "description": "No such resource",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "406": {
                  "description": "Content type was not supported",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "409": {
                  "description": "Resource state conflicts",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "500": {
                  "description": "Server error",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               },
               "default": {
                  "description": "Unexpected error",
                  "schema": {
                     "$ref": "#/definitions/ErrorModel"
                  }
               }
            }
         }
      },
...
   },
   "definitions": {
...
      "VMInformation": {
         "type": "object",
         "required": [
            "id"
         ],
         "properties": {
            "id": {
               "type": "string"
            },
            "cpu": {
               "$ref": "#/definitions/VMCPU"
            },
            "memory": {
               "$ref": "#/definitions/VMMemory"
            }
         }
      },
...
      "VMCloneParameter": {
         "type": "object",
         "required": [
            "name",
            "parentId"
         ],
         "properties": {
            "name": {
               "description": "New VM name",
               "type": "string"
            },
            "parentId": {
               "description": "Existing VM ID to clone.",
               "type": "string"
            }
         }
      },
      "VMID": {
         "required": [
            "id",
            "path"
         ],
         "properties": {
            "id": {
               "type": "string",
               "description": "ID of the VM"
            },
            "path": {
               "type": "string",
               "description": "Path of the VM"
            }
         }
      },
      "VMPowerState": {
         "type": "object",
         "required": [
            "power_state"
         ],
         "properties": {
            "power_state": {
               "type": "string",
               "enum": [
                  "poweredOn",
                  "poweredOff",
                  "paused",
                  "suspended"
               ]
            }
         }
      },
      "VMCPU": {
         "description": "The CPU information of VM",
         "type": "object",
         "readOnly": true,
         "properties": {
            "processors": {
               "$ref": "#/definitions/VMProcessors"
            }
         }
      },
      "VMProcessors": {
         "description": "Number of processor cores",
         "type": "integer",
         "format": "int32",
         "readOnly": true,
         "default": 1
      },
      "VMMemory": {
         "description": "Memory size in mega bytes",
         "type": "integer",
         "format": "int32",
         "readOnly": true,
         "default": 512
      },
      "VMPowerOperation": {
         "type": "string",
         "enum": [
            "on",
            "off",
            "shutdown",
            "suspend",
            "pause",
            "unpause"
         ]
      },
...
      "ErrorModel": {
         "required": [
            "code",
            "message"
         ],
         "properties": {
            "code": {
               "type": "integer",
               "format": "int32"
            },
            "message": {
               "type": "string"
            }
         }
      }
   }
}